+++ lib/braintree/transaction.rb
@@ -82,6 +82,7 @@ module Braintree
       end
     end
 
+    attr_reader :account_funding_transaction
     attr_reader :ach_return_code
     attr_reader :ach_return_responses
     attr_reader :acquirer_reference_number
@@ -171,6 +172,7 @@ module Braintree
     attr_reader :tax_exempt
     attr_reader :three_d_secure_info
     attr_reader :type
+    attr_reader :upcoming_retry_date
     attr_reader :updated_at
     attr_reader :us_bank_account_details
     attr_reader :venmo_account_details