+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.30.0
+
+- Add `upcoming_retry_date` to `Transaction`
+- Add `remaining_file_evidence_storage` to Dispute
+- Add `transfer_type` to Transaction
+- Add `transaction_retried` webhook
+
 ## 4.29.0
 
 - Add Session Id to Customer Recommendations Payload